+++ test/clj/metosin/postgres/types_test.clj
@@ -7,7 +7,7 @@
 (deftest pgobject->clj-test
   (testing "json"
     (is (= {:foo "a"} (pgobject->clj (->PGobject "json" "{\"foo\":\"a\"}"))))
-    (is (= {:foo "a"} (pgobject->clj (jdbc/sql-value {:foo "a"}))))
+    (is (= {:foo "a"} (pgobject->clj (write-json {:foo "a"}))))
     (is (= ["a" "b"] (pgobject->clj (->PGobject "json" "[\"a\",\"b\"]"))))
-    (is (= ["a" "b"] (pgobject->clj (jdbc/sql-value ["a" "b"]))))
+    (is (= ["a" "b"] (pgobject->clj (write-json ["a" "b"]))))
     ))