+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Add some type hints to `metosin.dates`.
 - Remove dangerous keyword -> PGobject `ISQLValue` implementation from `metosin.postgres.types`
+- Removed `ISQLValue` implementation to write maps and vectors as JSON, instead provide new
+`write-json` function to create JSON PGobject.
 
 ## 0.3.0 (11.1.2016)